Model updates (#52)
* update django

* discord_message leaderboard on admin

More descriptive

* cleanup admin round health percentage

* add discordmessage fields

To track creation and updates to discord messages
Add name for easier findability

* remove discord id old

linked all existing icleargroups to new or existing discordmessage

* add id to admin

* add use_for_icg_duration bool to encounter

And default bool fields to False

* remove leaderboard_instance_group

* fill duration_encounters on create

When InstanceClearGroup is created it will fill the encounters that should be included in the leaderboard with it once.
Required some database edits to fix historical records.

* add use_in_leaderboard to DpsLog

with this we can remove practice runs of cerus cm from the total clear time

* leaderboard full clear respects new wings

When new wing is added duration_encounters will change. The leaderboard will only show the top times for the latest combination of duration_encounters

* fix clear duration rank

* add favicon

* fix strike leaderboard

* add samples count to rank emote

* remove old models